refactor: extract JavaTypePrimitiveMapping and JavaTypeMethodLookup from JavaType
Decompose JavaType.java (600 → 438 lines) by extracting two package-private
helper classes:

- JavaTypePrimitiveMapping: primitive-to-wrapper map, isWrapperType(),
  getWrapperTypeIfNecessary()
- JavaTypeMethodLookup: isMask(), trimLast(), getNextShorterInChain(),
  handleMthd(), and new buildMethodList() entry point

JavaType retains thin delegates for isWrapperType (public) and
getWrapperTypeIfNecessary (package-private) to preserve the existing API.
The methods Lazy body now delegates to JavaTypeMethodLookup.buildMethodList().

Removes dead isNotMask comment and 4 unused imports.

All 671 existing tests pass. 13 new characterization/contract tests pass.

Closes #654
