+++ core/ast/src/main/java/org/lgna/project/ast/JavaType.java
@@ -52,17 +52,12 @@ import edu.cmu.cs.dennisc.pattern.Lazy;
 import edu.cmu.cs.dennisc.property.PropertyUtilities;
 import edu.cmu.cs.dennisc.property.StringProperty;
 import org.lgna.project.annotations.*;
-import org.lgna.project.reflect.ClassInfoManager;
-import org.lgna.project.reflect.MethodInfo;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.function.BinaryOperator;
 
 /**
@@ -89,45 +84,13 @@ public class JavaType extends AbstractType<JavaConstructor, JavaMethod, JavaFiel
   public static final JavaType OBJECT_TYPE = getInstance(Object.class);
   public static final JavaType STRING_TYPE = getInstance(String.class);
 
-  private static Map<JavaType, JavaType> mapPrimitiveToWrapper = Maps.newHashMap();
-
-    static {
-    addPrimitiveToWrapper(Void.TYPE, Void.class);
-    addPrimitiveToWrapper(Boolean.TYPE, Boolean.class);
-    addPrimitiveToWrapper(Byte.TYPE, Byte.class);
-    addPrimitiveToWrapper(Character.TYPE, Character.class);
-    addPrimitiveToWrapper(Short.TYPE, Short.class);
-    addPrimitiveToWrapper(Integer.TYPE, Integer.class);
-    addPrimitiveToWrapper(Long.TYPE, Long.class);
-    addPrimitiveToWrapper(Float.TYPE, Float.class);
-    addPrimitiveToWrapper(Double.TYPE, Double.class);
-  }
-
-  private static void addPrimitiveToWrapper(Class<?> primitiveCls, Class<?> wrapperCls) {
-    mapPrimitiveToWrapper.put(JavaType.getInstance(primitiveCls), JavaType.getInstance(wrapperCls));
-  }
-
   public static boolean isWrapperType(AbstractType<?, ?, ?> type) {
-    return mapPrimitiveToWrapper.containsValue(type);
+    return JavaTypePrimitiveMapping.isWrapperType(type);
   }
 
   /* package-private */
   static AbstractType<?, ?, ?> getWrapperTypeIfNecessary(AbstractType<?, ?, ?> type) {
-    if (type instanceof JavaType javaType) {
-      if (javaType.isPrimitive()) {
-        JavaType wrapperType = mapPrimitiveToWrapper.get(javaType);
-        if (wrapperType != null) {
-          return wrapperType;
-        } else {
-          Logger.severe(type);
-          return type;
-        }
-      } else {
-        return type;
-      }
-    } else {
-      return type;
-    }
+    return JavaTypePrimitiveMapping.getWrapperTypeIfNecessary(type);
   }
 
   @Override
@@ -179,14 +142,6 @@ public class JavaType extends AbstractType<JavaConstructor, JavaMethod, JavaFiel
     }
     return rv;
   }
-
-  private static boolean isMask(int modifiers, int required) {
-    return (modifiers & required) != 0;
-  }
-
-  //  private static boolean isNotMask( int modifiers, int prohibited ) {
-  //    return (modifiers & prohibited) == 0;
-  //  }
 
   private JavaType(ClassReflectionProxy classReflectionProxy) {
     this.classReflectionProxy = classReflectionProxy;
@@ -302,33 +257,6 @@ public class JavaType extends AbstractType<JavaConstructor, JavaMethod, JavaFiel
     return this.getterSetterPairs.get();
   }
 
-  private static Class<?>[] trimLast(Class<?>[] src) {
-    Class<?>[] rv = new Class<?>[src.length - 1];
-    System.arraycopy(src, 0, rv, 0, rv.length);
-    return rv;
-  }
-
-  private static java.lang.reflect.Method getNextShorterInChain(java.lang.reflect.Method src) {
-    java.lang.reflect.Method rv;
-    Class<?> srcReturnCls = src.getReturnType();
-    String name = src.getName();
-    Class<?>[] srcParameterClses = src.getParameterTypes();
-    if (srcParameterClses.length > 0) {
-      Class<?>[] dstParameterClses = trimLast(srcParameterClses);
-      try {
-        rv = src.getDeclaringClass().getMethod(name, dstParameterClses);
-        if (rv.getReturnType() != srcReturnCls) {
-          rv = null;
-        }
-      } catch (NoSuchMethodException nsme) {
-        rv = null;
-      }
-    } else {
-      rv = null;
-    }
-    return rv;
-  }
-
   public ClassReflectionProxy getClassReflectionProxy() {
     return this.classReflectionProxy;
   }
@@ -438,39 +366,6 @@ public class JavaType extends AbstractType<JavaConstructor, JavaMethod, JavaFiel
     }
   }
 
-  private static void handleMthd(java.lang.reflect.Method mthd, List<JavaMethod> methods) {
-    int modifiers = mthd.getModifiers();
-    if (isMask(modifiers, Modifier.PUBLIC) || isMask(modifiers, Modifier.PROTECTED)) {
-      JavaMethod methodDeclaredInJava = JavaMethod.getInstance(mthd);
-      if (mthd.isAnnotationPresent(MethodTemplate.class)) {
-        MethodTemplate methodTemplate = mthd.getAnnotation(MethodTemplate.class);
-        if ((methodTemplate.visibility() == Visibility.PRIME_TIME) && (methodTemplate.isFollowedByLongerMethod() == false)) {
-          JavaMethod longer = methodDeclaredInJava;
-          java.lang.reflect.Method _mthd = mthd;
-          while (true) {
-            _mthd = getNextShorterInChain(_mthd);
-            if (_mthd != null) {
-              JavaMethod shorter = JavaMethod.getInstance(_mthd);
-              if (_mthd.isAnnotationPresent(MethodTemplate.class)) {
-                MethodTemplate shorterMethodTemplate = _mthd.getAnnotation(MethodTemplate.class);
-                if (shorterMethodTemplate.isFollowedByLongerMethod()) {
-                  longer.setNextShorterInChain(shorter);
-                  shorter.setNextLongerInChain(longer);
-                  longer = shorter;
-                } else {
-                  break;
-                }
-              }
-            } else {
-              break;
-            }
-          }
-        }
-      }
-      methods.add(methodDeclaredInJava);
-    }
-  }
-
   private final ClassReflectionProxy classReflectionProxy;
   private final Lazy<List<JavaConstructor>> constructors = new Lazy<List<JavaConstructor>>() {
     @Override
@@ -490,64 +385,7 @@ public class JavaType extends AbstractType<JavaConstructor, JavaMethod, JavaFiel
   private final Lazy<List<JavaMethod>> methods = new Lazy<List<JavaMethod>>() {
     @Override
     protected List<JavaMethod> create() {
-      Class<?> cls = classReflectionProxy.getReification();
-      if (cls != null) {
-        List<JavaMethod> methods = Lists.newLinkedList();
-        Set<java.lang.reflect.Method> methodSet = null;
-        Iterable<MethodInfo> methodInfos = ClassInfoManager.getMethodInfos(cls);
-        if (methodInfos != null) {
-          methodSet = new HashSet<java.lang.reflect.Method>();
-          for (MethodInfo methodInfo : methodInfos) {
-            try {
-              java.lang.reflect.Method mthd = methodInfo.getMthd();
-              if (mthd != null) {
-                handleMthd(mthd, methods);
-                methodSet.add(mthd);
-              }
-            } catch (RuntimeException re) {
-              //edu.cmu.cs.dennisc.print.PrintUtilities.println( "no such method", methodInfo, "on ", cls );
-              //re.printStackTrace();
-            }
-          }
-        }
-        for (java.lang.reflect.Method mthd : cls.getDeclaredMethods()) {
-          if ((methodSet != null) && methodSet.contains(mthd)) {
-            //pass
-          } else {
-            handleMthd(mthd, methods);
-          }
-        }
-
-        //update value templates for setters
-        for (JavaMethod method : methods) {
-          java.lang.reflect.Method mthd = method.getMethodReflectionProxy().getReification();
-          GetterTemplate propertyGetterTemplate = mthd.getAnnotation(GetterTemplate.class);
-          if (propertyGetterTemplate != null) {
-            java.lang.reflect.Method sttr = PropertyUtilities.getSetterForGetter(mthd);
-            JavaMethod setter;
-            if (sttr != null) {
-              setter = (JavaMethod) JavaMethod.getInstance(sttr).getLongestInChain();
-            } else {
-              setter = null;
-            }
-            if (setter != null) {
-              ValueTemplate valueTemplate = mthd.getAnnotation(ValueTemplate.class);
-              if (valueTemplate != null) {
-                JavaMethod m = setter;
-                while (m != null) {
-                  JavaMethodParameter parameter0 = (JavaMethodParameter) m.getRequiredParameters().getFirst();
-                  parameter0.setValueTemplate(valueTemplate);
-                  m = m.getNextShorterInChain();
-                }
-              }
-            }
-          }
-        }
-
-        return Collections.unmodifiableList(methods);
-      } else {
-        return Collections.emptyList();
-      }
+      return JavaTypeMethodLookup.buildMethodList(classReflectionProxy.getReification());
     }
   };
   private final Lazy<List<JavaField>> fields = new Lazy<List<JavaField>>() {
